--- a/blendsql/db/__init__.py
+++ b/blendsql/db/__init__.py
@@ -1,5 +1,5 @@
 from ._database import Database
 from ._sqlite import SQLite
 from ._postgres import PostgreSQL
-from ._pandas import Pandas
+from ._duckdb import DuckDB
 from .utils import single_quote_escape, double_quote_escape