--- a/blendsql/db/__init__.py
+++ b/blendsql/db/__init__.py
@@ -2,4 +2,5 @@ from ._database import Database
 from ._sqlite import SQLite
 from ._postgres import PostgreSQL
 from ._duckdb import DuckDB
+from ._pandas import Pandas
 from .utils import single_quote_escape, double_quote_escape