--- a/blendsql/db/_database.py
+++ b/blendsql/db/_database.py
@@ -1,4 +1,5 @@
-from typing import Generator, Union, List, Collection, Type, Optional
+from typing import Generator, Union, List, Collection, Optional
+from typing_extensions import Callable
 import pandas as pd
 from attr import attrib
 from sqlalchemy.engine import URL
@@ -82,7 +83,7 @@ class Database(ABC):
 
     @abstractmethod
     def execute_to_list(
-        self, query: str, to_type: Optional[Type] = lambda x: x
+        self, query: str, to_type: Optional[Callable] = lambda x: x
     ) -> list:
         """A lower-level execute method that doesn't use the pandas processing logic.
         Returns results as a list.