- Added `DuckDB` class, with classmethods `from_pandas` and `from_sqlite`
    - Also added `Pandas` class, which is just a wrapper over the `from_pandas` method
- Parametrized `test_single_table_blendsql.py` and  `test_multi_table_blendsql.py` to use both SQLite and new DuckDB databases (245e886002f9b49aa6c9e2cf1cf7558158767a5a)
- `get_sqlglot_schema` -> cached property `sqlglot_schema` (7e0cbf4a55f0cb55c94aa919efa05c2c0aecf943, d618bbc6807f1e37dce39e790f483e2c9150ddbd)
- Mypy type hinting (018a33ae8f7da8323ae6b4409490ad8acf04064b)
- Documentation updates


## Benchmark Runtimes

The benchmarks saw no noticeable change in performance with this PR.

Before:
| Task           |   Average Runtime |   # Unique Queries |
|:---------------|------------------:|-------------------:|
| financials     |         0.0467936 |                  7 |
| rugby          |         0.267355  |                  4 |
| 1966_nba_draft |         0.113532  |                  2 |

After:
| Task           |   Average Runtime |   # Unique Queries |
|:---------------|------------------:|-------------------:|
| financials     |         0.0414736 |                  7 |
| rugby          |         0.255618  |                  4 |
| 1966_nba_draft |         0.124267  |                  2 |